Refactor file and package names for clearer API documentation
